requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
accelerate>=0.27.0
Pillow>=10.0.0
sentencepiece>=0.1.99